unify tools and add abstract canvas
