+++ python/packages/autogen-ext/src/autogen_ext/memory/canvas/__init__.py
@@ -1,4 +1,4 @@
-from ._canvas import Canvas
-from ._text_canvas import TextCanvasMemory
+from ._text_canvas import Canvas
+from ._text_canvas_memory import TextCanvasMemory
 
 __all__ = ["Canvas", "TextCanvasMemory"]
