+++ python/packages/autogen-ext/src/autogen_ext/memory/canvas/_text_canvas.py
@@ -1,62 +1,109 @@
-from autogen_core.memory import (
-    Memory,
-    MemoryContent,
-    MemoryQueryResult,
-    UpdateContextResult,
-    MemoryMimeType,
-)
-from autogen_core.model_context import ChatCompletionContext
-from autogen_core.models import SystemMessage
+import difflib
+from typing import Dict, List
 
-from typing import Any
-from ._canvas import Canvas
+from ._canvas import BaseCanvas
 
 
-class TextCanvasMemory(Memory):
+class FileRevision:
+    """Tracks the history of one file's content."""
+
+    def __init__(self, content: str, revision: int) -> None:
+        self.content = content
+        self.revision = revision  # e.g. an integer, a timestamp, or git hash
+
+
+class TextCanvas(BaseCanvas):
     """
-    A memory implementation that uses a Canvas for storing file-like content.
-    Inserts the current state of the canvas into the ChatCompletionContext on each turn
+    A simple in-memory 'Canvas' that can store multiple named files, each with revision history.
+    Allows retrieving content, generating diffs, and applying patches.
     """
 
-    def __init__(self, canvas: Canvas):
-        super().__init__()
-        self.canvas = canvas
+    def __init__(self):
+        # For each file, maintain a list of FileRevision. The latest revision is the end of the list.
+        self._files: Dict[str, List[FileRevision]] = {}
 
-    async def update_context(self, model_context: ChatCompletionContext) -> UpdateContextResult:
+    def _get_latest_revision_index(self, filename: str) -> int:
+        return len(self._files.get(filename, [])) - 1
+
+    def get_latest_content(self, filename: str) -> str:
         """
-        Inject the entire canvas summary (or a selected subset) as reference data.
-        Here, we just put it into a system message, but you could customize.
+        Returns the latest version of the file content, or empty if not found.
         """
-        snapshot = self.canvas.get_all_contents_for_context()
-        if snapshot.strip():
-            msg = SystemMessage(content=snapshot)
-            await model_context.add_message(msg)
+        revs = self._files.get(filename, [])
+        if not revs:
+            return ""
+        return revs[-1].content
 
-            # Return it for debugging/logging
-            memory_content = MemoryContent(content=snapshot, mime_type=MemoryMimeType.TEXT)
-            return UpdateContextResult(memories=MemoryQueryResult(results=[memory_content]))
-
-        return UpdateContextResult(memories=MemoryQueryResult(results=[]))
-
-    async def query(self, query: str | MemoryContent, cancellation_token=None, **kwargs: Any) -> MemoryQueryResult:
+    def add_or_update_file(self, filename: str, new_content: str) -> None:
         """
-        Potentially search for matching filenames or file content.
-        This example returns empty.
+        Overwrites the file with new_content, increments the revision index by 1.
         """
-        return MemoryQueryResult(results=[])
+        if filename not in self._files:
+            self._files[filename] = [FileRevision(new_content, 1)]
+        else:
+            current_revs = self._files[filename]
+            last_rev = current_revs[-1].revision
+            current_revs.append(FileRevision(new_content, last_rev + 1))
 
-    async def add(self, content: MemoryContent, cancellation_token=None) -> None:
+    def get_diff(self, filename: str, from_revision: int, to_revision: int) -> str:
         """
-        Example usage: Possibly interpret content as a patch or direct file update.
-        Could also be done by a specialized "CanvasTool" instead.
+        Returns a unified diff (string) between two revisions of the same file.
         """
-        # NO-OP here, leaving actual changes to the CanvasTool
-        pass
+        revisions = self._files.get(filename, [])
+        # naive checks
+        if not revisions or to_revision <= 0 or from_revision <= 0:
+            return ""
+        from_content = ""
+        to_content = ""
+        for rev in revisions:
+            if rev.revision == from_revision:
+                from_content = rev.content
+            elif rev.revision == to_revision:
+                to_content = rev.content
 
-    async def clear(self) -> None:
-        # If you want to clear the entire canvas:
-        self.canvas.__init__()  # naive: re-init
-        pass
+        diff = difflib.unified_diff(
+            from_content.splitlines(keepends=True),
+            to_content.splitlines(keepends=True),
+            fromfile=f"{filename}@r{from_revision}",
+            tofile=f"{filename}@r{to_revision}",
+        )
+        return "".join(diff)
 
-    async def close(self) -> None:
-        pass
+    def apply_patch(self, filename: str, patch_text: str) -> None:
+        """
+        Applies a unified diff patch to the latest version of a file and saves as a new revision.
+        """
+        # We retrieve the latest content, apply patch line by line.
+        current_content = self.get_latest_content(filename)
+        patched_lines = []
+        # difflib.patch.diff_bytes or external libs can do more robust patching. This is
+        # a simplified approach to illustrate the concept:
+        patched = difflib.restore(patch_text.splitlines(keepends=True), 2)
+        # difflib.restore expects a certain format that matches a 'context diff' or 'unified diff'
+        # If you want a robust patch apply, consider `unidiff` library or a real VCS-based approach.
+        patched_lines = list(patched)
+
+        new_content = "".join(patched_lines)
+        self.add_or_update_file(filename, new_content)
+
+    def list_files(self) -> Dict[str, int]:
+        """
+        Returns a dict of filename -> latest revision number.
+        """
+        results = {}
+        for fname, revs in self._files.items():
+            if revs:
+                results[fname] = revs[-1].revision
+        return results
+
+    def get_all_contents_for_context(self) -> str:
+        """
+        Quick way to produce a text block summarizing the entire canvas and its latest revisions
+        for LLM consumption.
+        """
+        out = ["=== CANVAS FILES ==="]
+        for fname, revs in self._files.items():
+            latest_rev = revs[-1]
+            out.append(f"File: {fname} (rev {latest_rev.revision}):\n{latest_rev.content}\n")
+        out.append("=== END OF CANVAS ===")
+        return "\n".join(out)
