--- python/packages/autogen-ext/src/autogen_ext/tools/canvas/__init__.py
@@ -1,3 +1,0 @@
-from ._canvas_writer import UpdateFileTool, ApplyPatchTool
-
-__all__ = ["UpdateFileTool", "ApplyPatchTool"]
